requirements.txt
fastapi
uvicorn
pdf2image
pillow
numpy
opencv-python-headless
pytesseract
pulp
rapidfuzz
requests
python-multipart
pytest
